requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
certifi==2026.2.25
charset_normalizer==3.4.5
deep-translator==1.11.4
discord.py==2.7.1
frozenlist==1.8.0
google-auth==2.41.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
